moved isTouchingMaterial to GlowEntity and added a note
